requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
google-generativeai==0.3.1
Pillow==10.0.0
python-multipart==0.0.6
httpx==0.24.1
starlette==0.27.0
typing-extensions==4.7.1